pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "autogen-ext>=0.5.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyautogen>=0.8.6",
    "python-dotenv>=1.1.0",
]
